Introduce futures for import path interpretation
It's important for prefetchingthat we can defer the discovery process
of sources for import paths (which inevitably involves some network interaction)
into a background goroutine. At the same time, it's also crucial that we
can choose whether or not to parallelize such things from outside (the
caller). This futures implementation satisfies both requirements.