Also return string ident from sourceFuture
This helps resolve a problem that will probably only exist for vcs-type
sources, where there could be a difference between the input ident
(e.g., a plain import path) and the actual on-disk ident of the
resulting source (e.g., a full URL). The manager needs to know which
unique string ident is in use, because that will become an access path
for lookups (in addition to the input path).