Add default branch handling for hg
This actually has a number of additions/changes, including using
bookmark data from hg. Handling of default branches is an interplay
between bookmarks and branches - if the @ bookmark exists, then it's
taken as the default branch; otherwise, the 'default' branch (which is
itself a meaningful name in Mercurial) is marked as the default branch.