dep: write a basically correct manifest on init
init now scans the code tree of the project it is init-ing and attempts
to make a sane determination about what constraint it should apply for
each non-stdlib project root it encounters.

This adds another handwavy stub method - this one for figuring out what
version of code is currently on disk - but it illustrates intent, and we
can fix that later.