Refactor command processing and help text
Sorry for this big change. I think it's a big improvement.
First, I restructured how commands work.

- Commands are now structs that implement the command interface
- Flags are bound to parameters in the command struct
- Run is a method on the command struct, with access to flag vars

This allowed me to make a lot of consistency improvements.

- Good news: dep foo -v ... works
- Bad news: dep -v foo ... stopped working (but that was weird)
- Good news: dep foo -h works and looks consistent
- Bad news: dep help foo doesn't exist anymore (but that was weird)

With all of that in place, I was able to audit and refactor all of the
help texts. I think they're a lot better. The biggest individual change
was that I moved all of the ensure usage examples to a special flag,
ensure -examples. I hope that's acceptable.

If I screwed anything else up, please accept my apologies.