Section headers in hash inputs; tests refactor
To further improve debugging of issues with the input hashing, this adds
"section headers" - strings that are output prior to each type of data
that's present in the cache.

Also partially switched to progressive mutation table-based tests for
input hashing, and added test cases that cover salient combinations of
overrides, imports, and constraints.