Add funcs for typed constraint string output
These solve the problem, at least in the hasher, of the possibility for
strings representing different types of versions to collide. For
example, prior to this change, a branch constraint named "foo" and a
version constraint named "foo" could cause the hasher to produce the
same hash, even though the two inputs would not have admitted the same
solution set.