Fix bad conditional in syncLocal
This has to have just been an earlier mistake. It is a little worrisome,
though, that the move to sync.Once makes it impossible to retry the
local sync if the first one fails. Not because it's necessarily the
wrong behavior, but because it's inconsistent with how other parts of
the API behave.