Ensure writing deptree works for bzr, hg
The main fix here is avoiding creating an empty directory for the
destination, as copyDir() doesn't like that. Instead, we create only
up to the parent dir.

The other bit is ensuring the source repos exist in the cache before
attempting to export them, for both bzr and hg.

Addresses golang/dep#144