Modify the file writing process
When write the contents of `Manifest` and `Lock` to the file, it is inefficient (needs more memory allocation) to write the contents to bytes.Buffer before writing them to a file.

Using bytes.Buffer can be omitted, so { `Manifest` , `Lock` }.Marshal() is modified not to use bytes.Buffer and renamed to `WriteTo` to implement `io.WriterTo` interface.