Non-existing import path with less side effects
Using a github.com repository for a fixture intended to not exist has
the unfortunate side effect of, at least sometimes, cauisng the tool to
hang when prompting the user for a password. Hopefully this'll be
improved by using a valid, deducible import path pointing to a
nonexistent domain instead.