Always remove temp dir after test in vcs_source_test
Some tests in vcs_source_test did not always delete the temp dir they
created. The git-source tests for example only removed the temp dirs for
certain failure conditions, but not when the test passed. This is fixed
by deferring the cleanup function right after creating the temp dirs for
each test.