Just copy the method set of SourceManager into sourceBridge
This makes the method sets redundant, but sourceBridge shouldn't
have a Release() method because it only ever deals with a single solve
run. We could have a private unexported interface covering the
intersection of the method sets, but then we'd lose out on godoc.
Repeating the methods is the least-bad of our options.