docs: Create a FAQ
Based on a quick chat with @sdboyer we're going to start with a FAQ in
which we can collect people's questions. Whenever there's a recurring
question around `dep` please send a PR to the FAQ. This will serve as a
source of data for what the actual documentation should include.

This is a first step in resolving #331.