Put the examples at the top instead
Injecting the comments at the bottom is suboptimal, as it means that
additional dependencies injected by later `dep ensure` calls (under
the new spec) will be separated from the original `init` set.

This seems slightly annoying now, but once we add a flag to allow
bypassing example injection, it'll be less of a problem.