Add constructors for feedback logs
The feedback system was originally designed for reporting feedback as GOPATH
was scanned, which didn’t work well for logging from the importers. This moves
more of the logic around what should be logged back onto the caller and simplifies
the ConstraintFeedback.LogFeedback function to print what it is given.