Check if project is missing from vendor during ensure
If the vendor directory already exists, and the lock file hasn't
changed, even though a project may be missing from the vendor
directory, dep ensure would not add it.

This change checks if we're not already writing the vendor directory
whether any of the projects are missing from the vendor directory
and if they are, ensures the writer writes out the vendor directory.

Fixes #883.