dep: add NewManifest and add prune options
This change updates the manifest to accept a new table: prune. This
table defines pruning options accepted by dep. By default, dep will
always prune nested vendor directories. Three additional options can
be toggled in the manifest:

  * non-go: Prune non-Go files (will keep LICENSE & COPYING files)
  * go-tests: Prune Go test files
  * unused-packages: Prune unused Go packages from dependencies.

The implementation of these flags is not part of this commit.

A constructor for dep.Manifest was also added.

Signed-off-by: Ibrahim AshShohail <ibra.sho@gmail.com>