Handle bazaar revisions by talking to the actual repo
String parsing was a little error prone, so we'd prefer to ask the
bazaar repository whether a particular string is a valid revision ID.
To do this, we need a very slightly custom version of
disambiguateRevision for *bzrSource.

This revision also refactors TestSourceManager_InferConstraint pretty
dramatically to let it work on repositories coming from different
VCSes.