Explain the status of the project
Following the discussion https://github.com/golang/dep/issues/1070#issuecomment-325352595
"Agreed, dep cannot currently be trusted to not make changes after the importers are run.
The issue which tracks that is #908.", it is clear that the maintainers of
the tool do not see it as a reliable project.

As such, the users should be correctly informed about the status of the
project and what are the goals in order to make it a production ready one.