fix(status): skip ignored pkgs in missing pkgs chk
This change skips the root packages from missing package check if
there's a path to the corresponding package in manifest's ignored list.

`digestMismatch` is removed from `runStatusAll()`. When there's
a digest mismatch, it's an error. And we already log to stderr but never
returned an actual error. `errInputDigestMismatch` is now returned as error
when there's a digest mismatch.

Some refactoring in how we handle the returned info from `runStatusAll()`
and show the errors and their solutions.