dep: Use TrimHiddenPackages in a standard way
This consolidates calls to parse the root project's code tree into a
single place, so as to avoid confusion about the standard, expected way
that dep's root project tree is to be created. It may not be fully
suitable for all status cases, but standardization is crucial for init
and ensure.