gps: Adaptively clean dirty git repos
Instead of forcing the user to clean up dirty git repositories, we can
take at least basic steps to doing it ourselves - or, if we detect
problems, instructing the user to fix it.

The overhead introduced here is a `git status` call, which will be
non-negligible on larger repos, but it's probably worth it for the
resilient behavior.