gps: Tests for adaptive git repo recovery/cleanup
This covers three basic cases - untracked files, modified files, and
some corruption in the .git directory. The first two are plausible; the
third is less so, as we don't know much about what real git failure
patterns could look like in the .git directory itself. However, it's
an adequate test inasmuch as it triggers failure in the basic git calls
we make, thereby triggering the recovery procedure.