Updated disjointConstraintFailure Error() to also display VersionType
Fixes #1343

If you have a matching version string (say "master" and "master") but they are of different VersionTypes (say "branch" and "revision"), you will get a really confusing error message.

This adds a String() method for version type and uses it to improve disjointConstraintFailure Error() to also show the VersionType so that if you hit the above you will get an actionable error message.