status: Add projectConstraint & constraintsCollection
The resultant collection obtained from collectConstraints() should
contain data about which project set a particular constraint on a
dependency project. `projectConstraint` consists of the project and the
constraint it sets on a given dependency.

And `constraintsCollection` is a map to store the pair of dependency project
and a collection of `projectConstraint` for different constraints.

This would be used by project argument status (status for a single project),
to show the importers of the project.