gps: Insulate against .git as file
`git ls-remote` was being run without a current working directory, which
- for some versions of git - has the unfortunate side effect of still
looking for a local .git in the cwd hierarchy, and failing if that .git
is a file instead of a directory (as it might be if the user is working
beneath a git submodule).

This relocates the call to somewhere safer, under gps' control, so that
we've a better (though not foolproof) guarantee of not colliding with a
.git file.

Fixes #1338.