status: use unpaired ver constr when checking updateability
This changes fixes a bug due to which the LATEST field in status was
missing for projects that were not in manifest but in lock and had
pairedVersion.

Since we check `bs.Version` to not be nil before entering into
updateability block, we can use the same value as the constraint for
matching with the version list obtained for the projects.