Update existing (pre-PR) tests to pass
One test relied upon constraints NOT being applied transitively.
The other is commented out because it was broken by supporting
transitive constraints without detecting constraints applied to a
package that is not a descendent. This test must be fixed in the
follow-up PR to support detecting and ignoring stale constraints.