status: LATEST should be of the same type as VERSION
This change modifies the LATEST field in status from always being a
revision to the same type as of the VERSION field.

If VERSION is a semver version, the LATEST would have the latest semver
version of the project constrained by the effective constraint.
If VERSION is a branch version, the LATEST would have the latest
revision of that branch.