gps: Disable certain bzr tests on Windows
The precise underlying causes of the problem with bzr on Windows is
unclear - it could be just about bzr with Windows, or the version of
bzr, or some underlying system interaction. But it's causing appveyor
tests to erroneously fail, and bzr is so little-used that it's
acceptable to just skip the tests, for now.