prune: Introduce CascadingPruneOptions
This new type defers computation of the actual prune value for a given
project until the method is called, rather than trying to precompute it.
By deferring computation, we retain full fidelity on the original
cascading inputs.

Still a WIP hack - need to clean up, and make everything actually use this.