Implement vcs similar to go get
For private repositories and git over ssh repositories implementation of
dep was different from `go get`.

Examples:

  go get example.com/repo.git

Will use remote

  (scheme)://example.com/repo

The most important part that suffix .git tells go get to access the path
as git, but it drops the suffix to actually access the remote.
Go get will clone this repo as

  git clone git+ssh://example.com/repo $GOPATH/src/example.com/repo.git

It is very unfortunate that you will end up with .git suffixes but at least it
works, this is the only way to make `go get` to recognize this import as
git.

But in case of dep

  import "example.com/repo.git"

Will result for dep to try to access remote as

  (scheme)://example.com/repo.git

Which makes it impossible to teach go get to work with dep.

This change is making consistent behavior between *go get* and *dep* by
always dropping suffix *.git* for remotes (which is also not ideal, but
at least it works now with *go dep*).

Other changes related to the schemas, I made the order match to the
default schemes in *go get*