POC: external file-based registry
Projects building behind a corporate firewall need a way to deal with
local projects - and, potentially, locally-cached versions of upstream
projects.

Such builds may exist behind a strict firewall; additionally, it's not
always the case that the import path "example.com/xyz/bar" actually has
the go vanity metadata being served from it.

For these builds, we want a *local* way to intercept and redirect the
actions of dep, to get it to fetch from alternative locations - *without*
first attempting the HTTP(s) dereference of the importpath location.

We do this by introducing a _registry_. The only functionality that this
object must supply is that of the Deducer interface (which is now exported).

The example code only has one kind of registry to resolve these import paths.
It reads from the top-level file `Gopkg.reg, which should consist of lines
containing whitespace-separated elements. These are:

- the import path to match; this is a Golang regexp
- the registry root corresponding to that import path. This may contain
  references to capture groups in the import path pattern.
- one or more git URLs; these may also contain `$1`-style backreferences.

Example content:

```
^example\.com/xyz/([^/]*)(/.*)?$ example.com/xyz/$1 ssh://git@local-repo.org/foo/$1
```

Imports of "example.com/xyz/bar" will be cloned from "local-repo.org/foo/bar".

The current implementation is just a POC: its parsing of Gopkg.reg is
rudimentary. It only supports git repositories. However, the principle
is demonstrated: it requires minimal code-change to get the registry
behaviour into dep. Registry implementations that talk to alternative
repository APIs are, of course, possible.

Note: the behaviour specified in Gopkg.reg is *not* inherited by any
importing packages; it is a function solely of the local environment.