gps: export content of git submodules
The "git checkout-index" command is unaware of git submodules,
therefore their content was missed when exporting files.

With "git submodule foreach" the same operation (read-tree +
checkout-index) can also be done for all submodules, recursively. The
downside is that "foreach" invokes a shell command with certain
variables set. To achieve the desired functionality, we have to rely
on POSIX shell support.

OTOH, the command seems to get embedded by git in a larger shell
script because error messages when there are syntax errors show that the
command gets invoked via `eval`, so git already depends on a POSIX
shell.