dep: Encapsulate vendor verification into method
Some tests were complaining of a data race when writing to ctx.Out on a
vendor verification failure. It's not clear why those tests were
failing, but the complaint became a forcing function to refactor the
previously-sloppy use of a channel and os.Exit(1) within a goroutine to
encapsulation in a method on dep.Project.