testing: compare entire 'final' directory
So far, only the list of vendored projects and the top-level
Gopk.toml/lock files were compared. This way it was impossible to
write integration tests that also check the content of the vendored
projects.

Now a test case can specify that it's 'final' directory contains a
complete set of reference files, and then all files are checked
respectively updated (when --updated is given). Directories and file
meta still aren't checked.

The reference files of existing tests can be changed on a case-by-case
basis. This new comparison mode cannot be enabled unconditionally
because a few tests then fail because each test run creates
_vendor-<datetime> directories which are named differently in each
test run.