integration tests: run sequentially when updating test cases
When running "internal" and "external" versions of the same test case
in parallel, the test case data must be read-only to avoid race
conditions. When it is read/write because the tests were asked to
update the reference files with "-update", running in parallel must be
disabled.

This broke in practice when updating entire directory trees. The risk
of hitting the race is smaller when only updating single files, but it
is still a race.