add documentation explaining how to uninstall (#1954)
* add documentation explaining how to uninstall

There's good docs on how to install dep, but we're missing docs on how
to uninstall dep. This adds them. I thought about adding some text to
the top-level README.md file but figured uninstalling dep wasn't a vital
topic for top-level docs and that a separate doc in the docs/ tree was
more than enough.

Issue #1953

* Link in the uninstalling doc from the install doc page

* Aha, figured out how to add the uninstalling.md link in the docs site