verify: Exclude symlinks from hash digests
As we're moving in the direction of excluding symlinks entirely from
dependencies (hallelujah!), it's preferable to have the initial release
of vendor verification exclude symlinks from hash digests entirely.
For the most part, people shouldn't notice it, and we won't be forced to
bump the hash version when we stop writing out symlinks (though we may
bump it anyway).