Fix bug in preservation of vendor/.git
The previous code was attempting to copy the .git directory into
vendor/vendor/.git, i.e., with an extra level of nesting. It was
then failing to notice the error because it was only reporting errors of
type *os.LinkError.

Fix the bug by removing the layer of nesting and reporting all errors.

Fix #1997.