travis.yml, hack: fix tests on tip
The tip tests were failing due to the lack of a GOCACHE environment
variable, and bad defaults in tests where home directories did not
exist. Fix these by setting the GOCACHE to a place we know exists.

This and the Xenial fix combined allow us to re-enable the test suite,
e.g. to permit failures in the test suite to actually fail the build.

(We may still want to address the GOCACHE tests locally for users that
do not have a GOCACHE environment variable set; as it stands, they
still fail.)