Bitbucket API SCM resolution
This change introduces a new SCM resolution mechanism for Bitbucket repos, to
hopefully avoid false-positives when reaching out to Bitbucket to get the repo's
metadata. Bitbucket is a weird code repository solution in that it supports both
`git` and `hg` (mercurial) for hosting repos. This makes checking out source
code from Bitbucket less straightforward than other places.

Previously the deducer tried to make educated guesses around which SCM was in
use based on a few things, largely hints from the package's import path. It also
had some fallbacks in place in case it got it wrong, but it did still cause some
issues in a few cases when it got it wrong.

This change introduces a minimal internal Bitbucket API client (`bitbucketapi`)
with the single purpose of reaching out to Bitbucket to get the SCM type for the
repository. This deprecates the need for the earlier heuristics, and should
result in a correct result every time. This client was written to be consumed as
a singleton (calling functions on the package) to avoid complicating the
`bitbucketDeducer` further.

The caller of this new API method now makes use of a context.WithTimeout
specifying a 60 second timeout for the calls to Bitbucket. This is a somewhat
arbitrary timeout value, but considering that we're only fetching the top-level
metadata about a Bitbucket repository it seems fairly safe.

The client has no caching of previously seen SCM results for a package, which
could result in more API calls than we'd like. One optimization would be to keep
a map of previously resolved packages, and to serve the result from there.

Tests are also implemented for the `bitbucketapi` package, hitting `97.0%` code
coverage.

Signed-off-by: Tim Heckman <t@heckman.io>