gps: pass correct arguments to "git clean"
Previously, the "-x" in

    git submodule foreach --recursive git clean -x

was being applied to the "git submodule" command not the "git clean"
command. git submodule foreach expects the inner command to be wrapped
in a single string, so pass git clean as a single string.

Verified this works as expected on the command line.

Originally submitted as #2164 by Daniele Rodina
<geaaru@sabayonlinux.org>, re-submitted by me to amend the commit and
merge immediately.

Fixes #2164.
Fixes #2168.